requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.0.0
Django==4.1.10
django-cors-headers==4.2.0
djongo==1.3.6
dnspython==2.4.2
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
pymongo==3.12.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
